+++ CHANGELOG.md
@@ -14,6 +14,12 @@ This driver uses semantic versioning:
 - A change in the major version (e.g. 1.Y.Z -> 2.0.0) indicates _breaking_
   changes that require changes in your code to upgrade.
 
+## [Unreleased]
+
+### Added
+
+- Implemented hot backup API (DE-576)
+
 ## [8.4.1] - 2023-09-15
 
 ### Fixed
@@ -1711,6 +1717,7 @@ For a detailed list of changes between pre-release versions of v7 see the
 
   Graph methods now only return the relevant part of the response body.
 
+[unreleased]: https://github.com/arangodb/arangojs/compare/v8.4.1...HEAD
 [8.4.1]: https://github.com/arangodb/arangojs/compare/v8.4.0...v8.4.1
 [8.4.0]: https://github.com/arangodb/arangojs/compare/v8.3.1...v8.4.0
 [8.3.1]: https://github.com/arangodb/arangojs/compare/v8.3.0...v8.3.1